+++ interpreter/javascript/eon/Any.h
@@ -3,8 +3,8 @@
 
 union e_Any {
     void* ptr;
-    intptr_t val;
+    unsigned long val;
 };
 
 union e_Any e_Any_ptr(void* ptr);
-union e_Any e_Any_val(intptr_t val);
+union e_Any e_Any_val(unsigned long val);
